:truck: renamed day view #179
